refactor(cameras): move cloud sync button into Xiaomi discovery section
- Remove sync button from Cameras.svelte header
- Add sync button to DiscoveryPanel Xiaomi section top
- Import xiaomiSync API directly in DiscoveryPanel
- Same toast feedback pattern preserved
